r/archiveloop: add ntp-wait before archival
When the Raspberry Pi boots, its clock is likely to be quite wrong. This
was causing several file uploads to fail when using rclone to upload to
Google Drive. These uploads failed because the SSL certificate presented
by Google Drive appears to be invalid (i.e., the system clock indicates
the certificate will not be valid for many years).

While there is a (somewhat arbitrary) sleep time of 20 seconds before
archiving is attempted, this has proven insufficient. Raspbian ships
with ntp-wait, a utility that returns only after NTP has stabilized the
system clock. Before attempting to archive, call ntp-wait. This fully
resolves the aforementioned problems.