r/archiveloop: toggle sentry mode during archiving
If sentry mode is active when archiving begins, the car's alarm is
triggered. With this change, sentry mode is disabled before archiving
begins, mitigating that problem.

Additionally, whether or not sentry mode was active before, it is
enabled for the duration of archiving. This ensures the USB port is
powered on with greater reliability than sending "wake up" API requests.

After archiving is completed, sentry mode is turned off. For now, it is
not re-enabled – in a future revision, TeslaUSB will optionally return
sentry mode to its pre-archive state (e.g., if sentry mode was enabled,
ensure it is enabled after archiving).